+++ package.json
@@ -10,6 +10,7 @@
   "license": "ISC",
   "dependencies": {
     "@slack/bolt": "^3.4.0",
+    "dotenv": "^10.0.0",
     "mongodb": "^3.6.9",
     "newsapi": "^2.4.1",
     "node-fetch": "^2.6.1"
